Move DataSource and associated types to the logmodel subproject
