+++ logmodel/src/main/java/name/mlopatkin/andlogview/logmodel/SourceMetadata.java
@@ -14,22 +14,12 @@
  * limitations under the License.
  */
 
-package name.mlopatkin.andlogview.liblogcat;
+package name.mlopatkin.andlogview.logmodel;
 
-public class SourceMetadataItem {
-    private final String displayName;
-    private final String value;
+import java.util.Collection;
 
-    public SourceMetadataItem(String displayName, String value) {
-        this.displayName = displayName;
-        this.value = value;
-    }
-
-    public String getDisplayName() {
-        return displayName;
-    }
-
-    public String getValue() {
-        return value;
-    }
+/** A collection of metadata that describes the data source. */
+public interface SourceMetadata {
+    /** @return the collection of metadata items available for the data source */
+    Collection<SourceMetadataItem> getMetadataItems();
 }
