+++ src/name/mlopatkin/andlogview/liblogcat/ddmlib/AdbDataSource.java
@@ -23,13 +23,13 @@ import name.mlopatkin.andlogview.device.DeviceChangeObserver;
 import name.mlopatkin.andlogview.device.DeviceGoneException;
 import name.mlopatkin.andlogview.device.OutputTarget;
 import name.mlopatkin.andlogview.device.ProvisionalDevice;
-import name.mlopatkin.andlogview.liblogcat.DataSource;
-import name.mlopatkin.andlogview.liblogcat.RecordListener;
-import name.mlopatkin.andlogview.liblogcat.SourceMetadata;
 import name.mlopatkin.andlogview.liblogcat.ddmlib.AdbBuffer.BufferReceiver;
+import name.mlopatkin.andlogview.logmodel.DataSource;
 import name.mlopatkin.andlogview.logmodel.Field;
 import name.mlopatkin.andlogview.logmodel.LogRecord;
 import name.mlopatkin.andlogview.logmodel.LogRecord.Buffer;
+import name.mlopatkin.andlogview.logmodel.RecordListener;
+import name.mlopatkin.andlogview.logmodel.SourceMetadata;
 import name.mlopatkin.andlogview.utils.events.Observable;
 import name.mlopatkin.andlogview.utils.events.ScopedObserver;
 import name.mlopatkin.andlogview.utils.events.Subject;
