+++ src/name/mlopatkin/andlogview/liblogcat/file/DumpstateFileDataSource.java
@@ -16,15 +16,15 @@
 package name.mlopatkin.andlogview.liblogcat.file;
 
 import name.mlopatkin.andlogview.config.Configuration;
-import name.mlopatkin.andlogview.liblogcat.DataSource;
 import name.mlopatkin.andlogview.liblogcat.LogRecordParser;
 import name.mlopatkin.andlogview.liblogcat.ProcessListParser;
-import name.mlopatkin.andlogview.liblogcat.RecordListener;
-import name.mlopatkin.andlogview.liblogcat.SourceMetadata;
 import name.mlopatkin.andlogview.liblogcat.file.ParsingStrategies.Strategy;
+import name.mlopatkin.andlogview.logmodel.DataSource;
 import name.mlopatkin.andlogview.logmodel.Field;
 import name.mlopatkin.andlogview.logmodel.LogRecord;
 import name.mlopatkin.andlogview.logmodel.LogRecord.Buffer;
+import name.mlopatkin.andlogview.logmodel.RecordListener;
+import name.mlopatkin.andlogview.logmodel.SourceMetadata;
 
 import com.google.common.base.CharMatcher;
