+++ test/name/mlopatkin/andlogview/liblogcat/file/FileDataSourceFactoryTest.java
@@ -18,7 +18,7 @@ package name.mlopatkin.andlogview.liblogcat.file;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 
-import name.mlopatkin.andlogview.liblogcat.DataSource;
+import name.mlopatkin.andlogview.logmodel.DataSource;
 import name.mlopatkin.andlogview.logmodel.Field;
 import name.mlopatkin.andlogview.logmodel.LogRecord.Buffer;
